+++ src/raftstore/store/worker/split_check.rs
@@ -236,139 +236,3 @@ fn new_split_region(region_id: u64, epoch: RegionEpoch, split_key: Vec<u8>) -> M
         callback: None,
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use std::sync::mpsc;
-    use std::sync::Arc;
-
-    use tempdir::TempDir;
-    use rocksdb::Writable;
-    use kvproto::metapb::Peer;
-    use rocksdb::{ColumnFamilyOptions, DBOptions};
-
-    use raftstore::coprocessor::Config;
-    use storage::ALL_CFS;
-    use util::rocksdb::{new_engine_opt, CFOptions};
-    use util::properties::SizePropertiesCollectorFactory;
-    use util::config::ReadableSize;
-    use super::*;
-
-    #[test]
-    fn test_split_check() {
-        let path = TempDir::new("test-raftstore").unwrap();
-        let path_str = path.path().to_str().unwrap();
-        let db_opts = DBOptions::new();
-        let mut cf_opts = ColumnFamilyOptions::new();
-        let f = Box::new(SizePropertiesCollectorFactory::default());
-        cf_opts.add_table_properties_collector_factory("tikv.size-collector", f);
-        let cfs_opts = ALL_CFS
-            .iter()
-            .map(|cf| CFOptions::new(cf, cf_opts.clone()))
-            .collect();
-        let engine = Arc::new(new_engine_opt(path_str, db_opts, cfs_opts).unwrap());
-
-        let mut region = Region::new();
-        region.set_id(1);
-        region.set_start_key(vec![]);
-        region.set_end_key(vec![]);
-        region.mut_peers().push(Peer::new());
-        region.mut_region_epoch().set_version(2);
-        region.mut_region_epoch().set_conf_ver(5);
-
-        let (tx, rx) = mpsc::sync_channel(100);
-        let ch = RetryableSendCh::new(tx, "test-split");
-        let mut cfg = Config::default();
-        cfg.region_max_size = ReadableSize(100);
-        cfg.region_split_size = ReadableSize(60);
-
-        let mut runnable = Runner::new(
-            engine.clone(),
-            ch.clone(),
-            Arc::new(CoprocessorHost::new(cfg, ch.clone())),
-        );
-
-        // so split key will be z0006
-        for i in 0..7 {
-            let s = keys::data_key(format!("{:04}", i).as_bytes());
-            engine.put(&s, &s).unwrap();
-        }
-
-        runnable.run(Task::new(&region));
-        // size has not reached the max_size 100 yet.
-        match rx.try_recv() {
-            Ok(Msg::ApproximateRegionSize { region_id, .. }) => {
-                assert_eq!(region_id, region.get_id());
-            }
-            others => panic!("expect recv empty, but got {:?}", others),
-        }
-
-        for i in 7..11 {
-            let s = keys::data_key(format!("{:04}", i).as_bytes());
-            engine.put(&s, &s).unwrap();
-        }
-
-        // Approximate size of memtable is inaccurate for small data,
-        // we flush it to SST so we can use the size properties instead.
-        engine.flush(true).unwrap();
-
-        runnable.run(Task::new(&region));
-        match rx.try_recv() {
-            Ok(Msg::ApproximateRegionSize { region_id, .. }) => {
-                assert_eq!(region_id, region.get_id());
-            }
-            others => panic!("expect approximate region size, but got {:?}", others),
-        }
-        match rx.try_recv() {
-            Ok(Msg::SplitRegion {
-                region_id,
-                region_epoch,
-                split_key,
-                ..
-            }) => {
-                assert_eq!(region_id, region.get_id());
-                assert_eq!(&region_epoch, region.get_region_epoch());
-                assert_eq!(split_key, b"0006");
-            }
-            others => panic!("expect split check result, but got {:?}", others),
-        }
-
-        // So split key will be z0003
-        for i in 0..6 {
-            let s = keys::data_key(format!("{:04}", i).as_bytes());
-            for cf in ALL_CFS {
-                let handle = engine.cf_handle(cf).unwrap();
-                engine.put_cf(handle, &s, &s).unwrap();
-            }
-        }
-        for cf in ALL_CFS {
-            let handle = engine.cf_handle(cf).unwrap();
-            engine.flush_cf(handle, true).unwrap();
-        }
-
-        runnable.run(Task::new(&region));
-        match rx.try_recv() {
-            Ok(Msg::ApproximateRegionSize { region_id, .. }) => {
-                assert_eq!(region_id, region.get_id());
-            }
-            others => panic!("expect approximate region size, but got {:?}", others),
-        }
-        match rx.try_recv() {
-            Ok(Msg::SplitRegion {
-                region_id,
-                region_epoch,
-                split_key,
-                ..
-            }) => {
-                assert_eq!(region_id, region.get_id());
-                assert_eq!(&region_epoch, region.get_region_epoch());
-                assert_eq!(split_key, b"0003");
-            }
-            others => panic!("expect split check result, but got {:?}", others),
-        }
-
-        drop(rx);
-        // It should be safe even the result can't be sent back.
-        runnable.run(Task::new(&region));
-    }
-}
